+++ open2fa/__init__.py
@@ -1,0 +1,3 @@
+from .version import __version__
+from .totp import generate_totp_2fa_code, TOTP2FACode
+from .main import Open2FA