+++ CHANGELOG.md
@@ -33,3 +33,11 @@ Also added set_version.py in repo root dir to update the version in open2fa/vers
 Added test for --version and -v flags
 
 Changed the way --version is handled in cli.py
+
+## 1.0.8
+
+Handeled empty cli args in cli.py by printing help message
+
+Added tests to verify this
+
+Updated parse_args to return the parser instead of the args, and updated cli.py main() to reflect this